+++ src/com/kaylerrenslow/armaDialogCreator/data/xml/StringTableXmlWriter.java
@@ -1,0 +1,17 @@
+package com.kaylerrenslow.armaDialogCreator.data.xml;
+
+import com.kaylerrenslow.armaDialogCreator.arma.stringtable.StringTable;
+import com.kaylerrenslow.armaDialogCreator.arma.stringtable.StringTableWriter;
+import org.jetbrains.annotations.NotNull;
+
+import java.io.IOException;
+
+/**
+ @author Kayler
+ @since 12/25/2016 */
+public class StringTableXmlWriter implements StringTableWriter {
+	@Override
+	public void writeTable(@NotNull StringTable table) throws IOException {
+
+	}
+}